requirements.txt
absl-py==0.12.0
aiohttp-cors==0.7.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
anyio==2.2.0
argon2-cffi==20.1.0
argparse==1.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
autopage==0.5.1
babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
bayesian-optimization==1.2.0
bleach==3.3.0
blessed==1.19.1
blessings==1.7
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.0
click==7.1.2
cliff==3.10.1
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.4
colorful==0.5.4
colorlog==6.6.0
configparser==5.0.2
cryptography==3.4.6
cycler==0.10.0
cytoolz==0.11.0
dask==2021.3.0
decorator==4.4.2
defusedxml==0.7.1
diffdist==0.1
distlib==0.3.4
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.3.0
entrypoints==0.3
fairscale==0.4.6
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.5
gitpython==3.1.14
google-api-core==2.8.2
google-auth-oauthlib==0.4.3
google-auth==1.28.0
googleapis-common-protos==1.56.3
gpustat==1.0.0b1
gql==0.2.0
graphql-core==1.1
greenlet==1.1.2
grpcio==1.36.1
h5py==2.10.0
higher==0.2.1
idna==2.10
imageio-ffmpeg==0.4.7
imageio==2.9.0
importlib-metadata==3.7.3
importlib-resources==5.8.0
ipykernel==5.5.0
ipypb==0.5.0
ipython-genutils==0.2.0
ipython==7.21.0
ipywidgets==7.7.1
jedi==0.18.0
jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonpatch==1.32
jsonpickle==1.5.2
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.1.1
jupyterlab==3.0.12
kiwisolver==1.3.1
kornia==0.6.5
lightning-bolts==0.5.0
llvmlite==0.38.1
mako==1.2.0
markdown==3.3.4
markupsafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mkl-fft==1.3.0
msgpack==1.0.4
multidict==6.0.2
munch==2.5.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.5
networkx==2.5
notebook==6.3.0
numba==0.55.2
numpy==1.19.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus-context==0.1.2
opencensus==0.9.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.1
pandocfilters==1.4.2
parso==0.8.1
pathtools==0.1.2
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.2.0
pip==21.0.1
platformdirs==2.5.2
plotly==4.9.0
prettytable==3.3.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==7.0.0
py-spy==0.3.12
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydeprecate==0.3.1
pygments==2.8.1
pyopenssl==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.1
pytz==2021.1
pywavelets==1.1.1
pyyaml==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
qtpy==2.1.0
requests-oauthlib==1.3.0
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
runstats==1.8.0
sacred==0.8.1
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.2
send2trash==1.5.0
sentry-sdk==1.0.0
setuptools
shortuuid==1.0.1
six==1.15.0
skorch==0.11.0
smart-open==6.0.0
smmap==3.0.5
sniffio==1.2.0
sqlalchemy==1.4.39
stevedore==3.5.0
subprocess32==3.5.4
tabulate==0.8.10
tensorboard-plugin-wit==1.8.0
tensorboard==2.2.0
tensorboardx==2.5.1
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
timm==0.4.5
toolz==0.11.1
tornado==6.1
tqdm==4.48.2
traitlets==5.0.5
typing-extensions==4.2.0
urllib3==1.26.4
virtualenv==20.15.0
visdom==0.1.8.9
wandb==0.10.23
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.6.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.4.1